refactor directory and structs
